+++ example/android/app/src/main/java/com/mz/td_sip_plugin_example/MainActivity.java
@@ -3,4 +3,5 @@ package com.mz.td_sip_plugin_example;
 import io.flutter.embedding.android.FlutterActivity;
 
 public class MainActivity extends FlutterActivity {
+
 }