+++ example/lib/main.dart
@@ -1,7 +1,9 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:td_sip_plugin/td_sip_plugin.dart';
-import 'package:td_sip_plugin/TDDisplayView.dart';
+import 'package:mz_back_plugin/mz_back_plugin.dart';
 import 'package:permission_handler/permission_handler.dart';
+import 'package:td_sip_plugin/TDDisplayView.dart';
+import 'package:td_sip_plugin/td_sip_plugin.dart';
 
 void main() {
   runApp(MyApp());
@@ -13,26 +15,95 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
+  @override
+  void initState() {
+    super.initState();
+
+    /// 初始化云对讲插件
+    TdSipPlugin.initial();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      initialRoute: "/",
+      routes: <String, WidgetBuilder>{
+        "/": (BuildContext context) => HomePage(),
+        "/td_sip_page": (BuildContext context) => SipPage(),
+      },
+    );
+  }
+}
+
+class HomePage extends StatefulWidget {
+  @override
+  _HomePageState createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage>
+    with WidgetsBindingObserver, TdSipObserver {
   String _loginStatus = "";
   String _callStatus = "";
+  bool _isPaused = false; //判断是都处于后台
 
   @override
   void initState() {
     super.initState();
     _getLoginStatus();
-    TdSipPlugin.addSipReceiver(onEvent: (event) {
-      print(event.toString());
-      if (event["eventName"] == "loginStatus") {
-        setState(() {
-          _loginStatus =
-              "${TDSipLoginStatus.values[event["loginStatus"]]}";
-        });
-      } else {
-        setState(() {
-          _callStatus = event["eventName"];
-        });
-      }
+
+    WidgetsBinding.instance.addObserver(this);
+    TdSipPlugin.addSipObserver(this);
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+
+    /// 移除监听
+    TdSipPlugin.removeSipObserver(this);
+    super.dispose();
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    super.didChangeAppLifecycleState(state);
+    _isPaused = state == AppLifecycleState.paused;
+  }
+
+  @override
+  void tdSipLoginStatus(TDSipLoginStatus status) {
+    super.tdSipLoginStatus(status);
+    setState(() {
+      _loginStatus = "$status";
     });
+  }
+
+  @override
+  void tdSipDidCallOut() {
+    super.tdSipDidCallOut();
+    Navigator.of(context).pushNamed("/td_sip_page");
+  }
+
+  @override
+  void tdSipDidCallEnd() {
+    super.tdSipDidCallEnd();
+    Navigator.of(context).pop();
+  }
+
+  @override
+  void tdSipDidReceiveCallForID(String sipID) {
+    super.tdSipDidReceiveCallForID(sipID);
+
+    /// 设置呼叫页面息屏显示后，只有iOS需要做页面跳转处理，Android已在原生底层处理，只需要实现路由为的页面即可
+    /// ⚠️ 路由"/td_sip_page"为固定的呼叫页面路由
+    if (defaultTargetPlatform == TargetPlatform.android && _isPaused) {
+      /// 这里可以本地存储相关呼叫信息，然后在SipPage里面去获取
+      /// 比如shared_preferences
+      TdSipPlugin.showSipPage();
+    } else {
+      Navigator.of(context)
+          .pushNamed("/td_sip_page", arguments: {"sipID": sipID});
+    }
   }
 
   void _getLoginStatus() async {
@@ -51,7 +122,7 @@ class _MyAppState extends State<MyApp> {
       ///用户点击了 拒绝且不再提示
     } else {
       PermissionStatus newStatus = await permission.request();
-      if(newStatus.isGranted) {
+      if (newStatus.isGranted) {
         TdSipPlugin.call("1907556514130605");
       }
     }
@@ -59,8 +130,8 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
+    return WillPopScope(
+      child: Scaffold(
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
@@ -78,12 +149,7 @@ class _MyAppState extends State<MyApp> {
                       sipID: "1100000004",
                       sipPassword: "2ebdeb8b65d320b2",
                       sipDomain: "47.106.186.8",
-                      sipPort: "8060",
-                      turnEnable: false,
-                      iceEnable: false,
-                      turnServer: "",
-                      turnUser: "",
-                      turnPassword: "");
+                      sipPort: "8060");
                 },
               ),
               ElevatedButton(
@@ -98,6 +164,88 @@ class _MyAppState extends State<MyApp> {
                   _checkPermission();
                 },
               ),
+              Text(
+                _callStatus,
+              ),
+            ],
+          ),
+        ),
+      ),
+      onWillPop: _onWillPop,
+    );
+  }
+
+  Future<bool> _onWillPop() {
+    if (defaultTargetPlatform == TargetPlatform.iOS) {
+      return Future.value(true);
+    }
+
+    /// 处理Android物理返回桌面后app销毁的问题
+    MzBackPlugin.navigateToSystemHome();
+    return Future.value(false);
+  }
+}
+
+class SipPage extends StatefulWidget {
+  @override
+  _SipPageState createState() => _SipPageState();
+}
+
+class _SipPageState extends State<SipPage> with TdSipObserver {
+  bool _showPlaceholder = true;
+
+  @override
+  void initState() {
+    super.initState();
+    TdSipPlugin.addSipObserver(this);
+  }
+
+  @override
+  void dispose() {
+    TdSipPlugin.removeSipObserver(this);
+    super.dispose();
+  }
+
+  @override
+  void tdSipStreamsDidBeginRunning() {
+    super.tdSipStreamsDidBeginRunning();
+    setState(() {
+      _showPlaceholder = false;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    print(ModalRoute.of(context).settings.arguments);
+    Size size = MediaQuery.of(context).size;
+    return Scaffold(
+      body: Container(
+        width: size.width,
+        height: size.height,
+        color: Colors.brown,
+        child: Center(
+          child: Column(
+            children: [
+              SizedBox(
+                height: 200,
+              ),
+              Container(
+                width: 200,
+                height: 120,
+                child: Stack(
+                  children: [
+                    TDDisplayView(),
+                    Visibility(
+                        visible: _showPlaceholder,
+                        child: Image.asset(
+                          "images/video.png",
+                          width: 200,
+                          height: 120,
+                          fit: BoxFit.cover,
+                        ))
+                  ],
+                ),
+              ),
               ElevatedButton(
                 child: Text("挂断"),
                 onPressed: () {
@@ -110,14 +258,6 @@ class _MyAppState extends State<MyApp> {
                   TdSipPlugin.answer();
                 },
               ),
-              Text(_callStatus,),
-              Container(
-                width: 200,
-                height: 120,
-                child: TDDisplayView(
-                  placeholder: "images/video.png",
-                ),
-              )
             ],
           ),
         ),