+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+* add the Full screen display of Android when the app is in background
+* remove the placeholder of TDDisplayView
+* fix some bug
+
 ## 0.0.3
 * fix the bug of error sipID when sip is coming